refactor: replace controller.ts to dir api
