+++ src/components/episodeList.ts
@@ -1,4 +1,4 @@
-import Controller from './controller';
+import Controller from './api/controller';
 import { EpisodesListItem } from './episodesListItem';
 
 export class EpisodeList {
